ecommerce-backend/requirements.txt
Django==4.2.3
djangorestframework==3.14.0
psycopg2==2.9.6
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.3
django-filter==23.2.0
django-cors-headers==3.14.0
djangorestframework-pagination==0.9.0